+++ CHANGELOG.md
@@ -8,6 +8,7 @@
     * Fixed a bug in `FormFlashFile::moveTo()` not deleting the old file
     * Fixed `FlexMediaTrait::getMedia()` trying to include uploaded but already moved media
     * Fixed `ImageMedium` constructor warning when file does not exist
+    * Fixed bad host header in PSR-7 (if using 'php -S localhost:8000 system/router.php')
 
 # v1.6.0-beta.8
 ## 01/25/2019